page query기능 분리
