+++ src/Main.java
@@ -1,3 +1,6 @@
+import WiseSayingController.WiseSayingController;
+import WiseSayingService.WiseSayingService;
+
 public class Main {
     public static void main(String[] args) {
         WiseSayingService service = new WiseSayingService();
